feat: Add placeholder pages and navigation wiring
This commit adds the base pages of information I believe will be
relevant - I'm treating the homepage as the one which will have all of
the registration details for the time being.

I currently have all of the class descriptions and much of the teachers'
bios, so I will shift now to adding all of the actual content to the
website before moving onwards with the (important) stylistic changes.