feat: Add pure logistical information to the homepage
This is an absolutely barebones representation of all of the relevant
information needed about the event. The extra pages will elaborate on
specifics, such as the teachers' bios and the class descriptions.

I'll fire the Netlify preview out to people to get feedback while I
desperately inhale food - the class descriptions are next, at which
point I'll pivot to making the website look nice.