feat: Add images of teachers, update text, hide empty pages
This commit clarifies certain details on the homepage and also
reinforces some of its information on the class page. It also removes
the venue and schedule page as although they are important, they aren't
strictly necessary for the initial release.

It also adds pictures to the teachers' page, which are not preprocessed
in any way yet, nor resized - the next big change will be to finally add
some structural CSS, to bring this to a coherent looking point.