feat: Add styling to tables, minor tweaks for readability
The tables are much easier to parse now since adding a border, which was
my biggest concern going forward for the content. I'm sufficiently happy
that everything is legible, so I will try to timebox the most effective
changes in the next 20 minutes I can before pushing the changes live.