feat: Add "Code of Conduct" page
This commit adds a page on the website for the code of conduct, which
also includes the e-mail details for the Safety Officer of the event.
Everyone had to read this code of conduct when signing up for the event:
this ensures there's another place to find and reference it, with the
third being at the desk during the event.