feat: Replace the entirety of the old website with a single landing page
This commit shifts the entire website for Emerald Swing 2023 into a
different folder, then creates a new single page website with some
slightly more simplified configuration to boot.

More information will be supplied when relevant, but this is most of the
things people we need to make a decision about whether or not to attend.
The next major update will be for aesthetic improvements.